+++ src/index.ts
@@ -1,4 +1,10 @@
-export { Viewer, Dimensions } from "./Viewer";
+export { Viewer } from "./Viewer";
+export { Dimensions } from "./dimensions";
 export { Element, ElementOptions } from "./Element";
 export { Space, Boundary, Coordinates } from "./Space";
-export { FunctionalTileLayer, TileFunction, TileCoordinates, GraphicsLayer } from "./layer";
+export {
+  FunctionalTileLayer,
+  TileFunction,
+  TileCoordinates,
+  GraphicsLayer
+} from "./layer";
