--- types/GraphicsLayer.d.ts
@@ -1,14 +1,0 @@
-/**
- * Enum for Optimaze graphics tile layers.
- */
-export declare enum GraphicsLayer {
-    BearingArea = 3,
-    RentableArea = 4,
-    GrossArea = 5,
-    Architect = 6,
-    Furniture = 7,
-    OuterWall = 8,
-    Elevators = 9,
-    Shafts = 11,
-    Atrium = 23,
-}
