+++ types/crs.d.ts
@@ -1,0 +1,3 @@
+import * as L from "leaflet";
+import { Dimensions } from "./dimensions";
+export declare function getCRS(dimensions: Dimensions): L.CRS;
